Fix windows build.
Got rid of all windows warnings.

Fixed a crash bug in ws_global_init when evdns_base_new fails (it would
free the ws_base_t before it freed the event base with in it.

+++ libws_compat.c
@@ -8,14 +8,14 @@ uint64_t libws_ntoh64(const uint64_t input)
     uint64_t rval;
     uint8_t *data = (uint8_t *)&rval;
 
-    data[0] = (input >> 56);
-    data[1] = (input >> 48);
-    data[2] = (input >> 40);
-    data[3] = (input >> 32);
-    data[4] = (input >> 24);
-    data[5] = (input >> 16);
-    data[6] = (input >> 8);
-    data[7] = (input >> 0);
+	data[0] = (uint8_t)(input >> 56);
+	data[1] = (uint8_t)(input >> 48);
+	data[2] = (uint8_t)(input >> 40);
+	data[3] = (uint8_t)(input >> 32);
+	data[4] = (uint8_t)(input >> 24);
+	data[5] = (uint8_t)(input >> 16);
+	data[6] = (uint8_t)(input >> 8);
+	data[7] = (uint8_t)(input >> 0);
 
     return rval;
 }
